Interactive add-in

Launches an interactive addin for insertion of roxygen2 comments in files. Allows selection of extra parameters for makeOxygen

interOxyAddIn()

Value

Nothing. Inserts roxygen2 comments in a file opened in the source editor.

Details

Open an .R file in Rstudio's source editor.

This addin requires shiny and miniUI to be installed (listed as Suggests in Description)

Launch the add-in via Addins -> interactiveOxygen or interOxyAddIn() in the console.

Add-in opens in the viewer panel.

Select function's/dataset's name in the source editor.

If objects cannot be found, the addin prompts to source the file.

Choose parameters for makeOxygen

Click Insert

Select next object's name

Rinse/Repeat

Click Quit when done with the file.

Author

Anton Grishin, Jonathan Sidi

Examples

if(interactive()) interOxyAddIn()