Return roxygen2 parameter calls from parameter dictionary

Return roxygen2 parameter calls from the intersection of the parameters listed in the package dictionary and the formals of a function

ls_param(obj, dictionary = "man-roxygen/Dictionary-1.R", print = TRUE)

Arguments

obj
function or name of function

dictionary
character, path_to_dictionary, Default: 'roxygen-man/Dictionary-1.R'

print
boolean print output to console, Default: TRUE

Value

character vector

Examples

repo='https://raw.githubusercontent.com/metrumresearchgroup/sinew/master/'
dict_loc=file.path(repo,'man-roxygen/Dictionary-1.R')
ls_param(sinew::makeOxygen,dictionary=dict_loc)

#> #' @param obj function or name of function
#> #' @param add_default boolean to add defaults values to the end of the PARAM 
#> #' @param add_fields character vector to add additional roxygen2 fields, 
#> #' @param use_dictionary character, path_to_dictionary, Default: NULL
#> #' @param print boolean print output to console, Default: TRUE