Inserts roxygen2 skeletons in file(s).

Applies makeOxygen function to all functions/dataframes in supplied file(s)

makeOxyFile(input = NULL, overwrite = FALSE, verbose = interactive(),
  ...)

Arguments

input
character, vector of path(s) to one or more .R files, a path to directory containing .R files, Default: NULL

overwrite
boolean, If TRUE overwrites file(s), FALSE writes "Oxy"- prefixed files in the same directory, Default: FALSE

verbose
boolean, If TRUE will print output to console and open edited files in the editor viewer, Defulat: interactive()

…
additional parameters passed to makeOxygen

Value

Nothing. Writes files with roxygen2 comments as a side effect

Details

If an object cannot be found it will be sourced into a temporary environment. If the file already contains roxygen2 comments they will be deleted to avoid duplication. Some functions may require attaching additional packages. For instance, if functions were defined with purrr's compose or partial functions, omission of purr:: in definitions will require library(purrr) before proceding with makeOxyFile.

See also

Examples

# NOT RUN {
if(interactive()){
 makeOxyFile() # default args, opens system file selection dialogue
 }
 makeOxyFile("./myRfunctions/utils.R") # on one R file
 makeOxyFile("./myRfunctions/") # on all R files in directory

 # adds more fields to defaults, passes "cut" to makeImport

 sinew_opts$append(list(add_fields=c("concept", "describeIn")))
 makeOxyFile("./myRfunctions/utils.R", cut = 5)

# }