Create lists of package exports

Useful for supplying packages to the force argument to pretty_namespace.

make_force_packages(packages)

Arguments

packages

(character) packages to include in list. When duplicate function names exist the order of packages determines which function will be selected first - IE the first package with the function name will include that function, the second package with the function name will not have it listed.

Value

(named list) with package names as names as all exports as a character vector

Examples

make_force_packages(c("utils"))
#> $utils
#>   [1] "write.socket"                         
#>   [2] "fixInNamespace"                       
#>   [3] "changedFiles"                         
#>   [4] "history"                              
#>   [5] "SweaveHooks"                          
#>   [6] "packageDate"                          
#>   [7] "installed.packages"                   
#>   [8] "citEntry"                             
#>   [9] "as.roman"                             
#>  [10] "suppressForeignCheck"                 
#>  [11] "aregexec"                             
#>  [12] "txtProgressBar"                       
#>  [13] "strcapture"                           
#>  [14] "xemacs"                               
#>  [15] "rtags"                                
#>  [16] "as.person"                            
#>  [17] "maintainer"                           
#>  [18] "aspell"                               
#>  [19] "bug.report"                           
#>  [20] "capture.output"                       
#>  [21] "remhash"                              
#>  [22] "assignInMyNamespace"                  
#>  [23] "aspell_package_vignettes"             
#>  [24] "memory.limit"                         
#>  [25] "person"                               
#>  [26] "ls.str"                               
#>  [27] "read.delim2"                          
#>  [28] "RweaveLatexWritedoc"                  
#>  [29] "hasName"                              
#>  [30] "getTxtProgressBar"                    
#>  [31] "news"                                 
#>  [32] "typhash"                              
#>  [33] "read.csv"                             
#>  [34] "RSiteSearch"                          
#>  [35] "debugger"                             
#>  [36] "demo"                                 
#>  [37] "isS3stdGeneric"                       
#>  [38] "Sweave"                               
#>  [39] "nsl"                                  
#>  [40] "modifyList"                           
#>  [41] "promptData"                           
#>  [42] "help.request"                         
#>  [43] "alarm"                                
#>  [44] "promptImport"                         
#>  [45] "close.socket"                         
#>  [46] "read.csv2"                            
#>  [47] "zip"                                  
#>  [48] "hsearch_db_concepts"                  
#>  [49] "assignInNamespace"                    
#>  [50] "askYesNo"                             
#>  [51] "read.delim"                           
#>  [52] "limitedLabels"                        
#>  [53] "is.relistable"                        
#>  [54] "toLatex"                              
#>  [55] "citFooter"                            
#>  [56] "head.matrix"                          
#>  [57] "View"                                 
#>  [58] "findMatches"                          
#>  [59] "create.post"                          
#>  [60] "globalVariables"                      
#>  [61] "vi"                                   
#>  [62] "bibentry"                             
#>  [63] "checkCRAN"                            
#>  [64] "Stangle"                              
#>  [65] "methods"                              
#>  [66] "sethash"                              
#>  [67] "type.convert"                         
#>  [68] "findLineNum"                          
#>  [69] "menu"                                 
#>  [70] "?"                                    
#>  [71] "aspell_package_C_files"               
#>  [72] "old.packages"                         
#>  [73] "hsearch_db_keywords"                  
#>  [74] "RShowDoc"                             
#>  [75] "browseURL"                            
#>  [76] "debugcall"                            
#>  [77] "is.hashtab"                           
#>  [78] "gethash"                              
#>  [79] "setTxtProgressBar"                    
#>  [80] "tail"                                 
#>  [81] "emacs"                                
#>  [82] "object.size"                          
#>  [83] "RweaveTryStop"                        
#>  [84] "undebugcall"                          
#>  [85] "download.file"                        
#>  [86] "loadhistory"                          
#>  [87] "RweaveChunkPrefix"                    
#>  [88] "summaryRprof"                         
#>  [89] "make.socket"                          
#>  [90] "xedit"                                
#>  [91] "data"                                 
#>  [92] "charClass"                            
#>  [93] "RtangleSetup"                         
#>  [94] "browseEnv"                            
#>  [95] "update.packages"                      
#>  [96] "untar"                                
#>  [97] "tail.matrix"                          
#>  [98] "getS3method"                          
#>  [99] "write.table"                          
#> [100] "strOptions"                           
#> [101] "rc.status"                            
#> [102] "asDateBuilt"                          
#> [103] "Rprofmem"                             
#> [104] "apropos"                              
#> [105] "getSrcDirectory"                      
#> [106] "as.relistable"                        
#> [107] "SweaveSyntaxNoweb"                    
#> [108] "SweaveSyntaxLatex"                    
#> [109] "citeNatbib"                           
#> [110] "read.fwf"                             
#> [111] "de.setup"                             
#> [112] "removeSource"                         
#> [113] "data.entry"                           
#> [114] "packageVersion"                       
#> [115] "as.personList"                        
#> [116] "citation"                             
#> [117] "citHeader"                            
#> [118] "read.fortran"                         
#> [119] "findCRANmirror"                       
#> [120] "aspell_package_R_files"               
#> [121] "RtangleFinish"                        
#> [122] "hashtab"                              
#> [123] "rc.getOption"                         
#> [124] "RtangleWritedoc"                      
#> [125] "readCitationFile"                     
#> [126] "stack"                                
#> [127] "package.skeleton"                     
#> [128] "memory.size"                          
#> [129] "de.ncols"                             
#> [130] "file.edit"                            
#> [131] "sessionInfo"                          
#> [132] "SweaveSyntConv"                       
#> [133] "RweaveLatex"                          
#> [134] "help"                                 
#> [135] "prompt"                               
#> [136] "getParseText"                         
#> [137] "osVersion"                            
#> [138] "url.show"                             
#> [139] "getCRANmirrors"                       
#> [140] "head"                                 
#> [141] "localeToCharset"                      
#> [142] "Rprof"                                
#> [143] "getParseData"                         
#> [144] "RweaveLatexSetup"                     
#> [145] "formatOL"                             
#> [146] "dump.frames"                          
#> [147] "read.table"                           
#> [148] "URLdecode"                            
#> [149] "numhash"                              
#> [150] "isS3method"                           
#> [151] "compareVersion"                       
#> [152] "count.fields"                         
#> [153] "tar"                                  
#> [154] "setRepositories"                      
#> [155] "edit"                                 
#> [156] "mirror2html"                          
#> [157] "relist"                               
#> [158] "chooseBioCmirror"                     
#> [159] "available.packages"                   
#> [160] "aspell_write_personal_dictionary_file"
#> [161] "getAnywhere"                          
#> [162] "hsearch_db"                           
#> [163] "personList"                           
#> [164] "fileSnapshot"                         
#> [165] "combn"                                
#> [166] "packageStatus"                        
#> [167] "read.socket"                          
#> [168] "find"                                 
#> [169] "RweaveLatexOptions"                   
#> [170] "page"                                 
#> [171] "write.csv"                            
#> [172] "dataentry"                            
#> [173] "getSrcFilename"                       
#> [174] "packageDescription"                   
#> [175] "file_test"                            
#> [176] "toBibtex"                             
#> [177] "browseVignettes"                      
#> [178] "setBreakpoint"                        
#> [179] "download.packages"                    
#> [180] "remove.packages"                      
#> [181] "RweaveLatexFinish"                    
#> [182] "chooseCRANmirror"                     
#> [183] "upgrade"                              
#> [184] "de"                                   
#> [185] "formatUL"                             
#> [186] "de.restore"                           
#> [187] "promptPackage"                        
#> [188] "write.csv2"                           
#> [189] "Rtangle"                              
#> [190] "packageName"                          
#> [191] "glob2rx"                              
#> [192] "unstack"                              
#> [193] "read.DIF"                             
#> [194] "savehistory"                          
#> [195] "vignette"                             
#> [196] "install.packages"                     
#> [197] "argsAnywhere"                         
#> [198] "help.search"                          
#> [199] "getSrcref"                            
#> [200] "makeRweaveLatexCodeRunner"            
#> [201] "RweaveEvalWithOpt"                    
#> [202] "lsf.str"                              
#> [203] "select.list"                          
#> [204] "cite"                                 
#> [205] "recover"                              
#> [206] "fix"                                  
#> [207] "new.packages"                         
#> [208] "aspell_package_Rd_files"              
#> [209] "str"                                  
#> [210] "RtangleRuncode"                       
#> [211] "process.events"                       
#> [212] "example"                              
#> [213] "warnErrList"                          
#> [214] "getSrcLocation"                       
#> [215] "flush.console"                        
#> [216] "maphash"                              
#> [217] "timestamp"                            
#> [218] "make.packages.html"                   
#> [219] "rc.options"                           
#> [220] "getFromNamespace"                     
#> [221] "clrhash"                              
#> [222] "adist"                                
#> [223] "contrib.url"                          
#> [224] "URLencode"                            
#> [225] "rc.settings"                          
#> [226] "help.start"                           
#> [227] "unzip"                                
#> [228] "pico"                                 
#>