Default and current sinew options

Options for functions in the sinew package. When running R code, the object sinew_opts (default options) is not modified by chunk headers (local chunk options are merged with default options), whereas sinew_opts_current (current options) changes with different chunk headers and it always reflects the options for the current chunk.

sinew_opts

sinew_opts_current

Format

An object of class list of length 5.

An object of class list of length 5.

Details

Normally we set up the global options once in the first code chunk in a document using sinew_opts$set(), so that all latter chunks will use these options. Note the global options set in one chunk will not affect the options in this chunk itself, and that is why we often need to set global options in a separate chunk.

A list of default chunk options, can be retrieved via sinew_opts$get()

Note

sinew_opts_current is read-only in the sense that it does nothing if you call sinew_opts_current$set(); you can only query the options via sinew_opts_current$get().

Examples

sinew_opts$get()
#> $add_fields
#> [1] "details"  "examples" "seealso"  "rdname"   "export"  
#> 
#> $cut
#> NULL
#> 
#> $pretty_print
#> [1] TRUE
#> 
#> $markdown_links
#> [1] FALSE
#> 
#> $author
#> [1] "AUTHOR [AUTHOR_2]"
#> 
#> $backref
#> [1] "src/filename.cpp"
#> 
#> $concept
#> [1] "CONCEPT_TERM_1 [CONCEPT_TERM_2]"
#> 
#> $describeIn
#> [1] "FUNCTION_NAME DESCRIPTION"
#> 
#> $details
#> [1] "DETAILS"
#> 
#> $example
#> [1] "path_to_file/relative/to/packge/root"
#> 
#> $examples
#> [1] "\n#' \\dontrun{\n#' if(interactive()){\n#'  #EXAMPLE1\n#'  }\n#' }"
#> 
#> $export
#> [1] ""
#> 
#> $family
#> [1] "FAMILY_TITLE"
#> 
#> $field
#> [1] "FIELD_IN_S4_RefClass DESCRIPTION"
#> 
#> $format
#> [1] "DATA_STRUCTURE"
#> 
#> $importClassesFrom
#> [1] "PKG CLASS_a [CLASS_b]"
#> 
#> $importMethodsFrom
#> [1] "PKG METHOD_a [METHOD_b]"
#> 
#> $include
#> [1] "FILENAME.R [FILENAME_b.R]"
#> 
#> $inherit
#> [1] "[PKG::]SOURCE_FUNCTION [FIELD_a FIELD_b]"
#> 
#> $inheritDotParams
#> [1] "[PKG::]SOURCE_FUNCTION"
#> 
#> $inheritSection
#> [1] "[PKG::]SOURCE_FUNCTION [SECTION_a SECTION_b]"
#> 
#> $keywords
#> [1] "KEYWORD_TERM"
#> 
#> $name
#> [1] "NAME"
#> 
#> $rdname
#> [1] "FUNCTION_NAME"
#> 
#> $references
#> [1] "BIB_CITATION"
#> 
#> $section
#> [1] "SECTION_NAME"
#> 
#> $source
#> [1] "\\url{http://somewhere.important.com/}"
#> 
#> $slot
#> [1] "SLOTNAME DESCRIPTION"
#> 
#> $template
#> [1] "FILENAME"
#> 
#> $templateVar
#> [1] "NAME VALUE"
#> 
#> $useDynLib
#> [1] "PKG [ROUTINE_a ROUTINE_b]"
#>